backend/requirements.txt
# Web framework
fastapi==0.115.5

# ASGI server for running FastAPI
uvicorn[standard]==0.32.1

# Form data handling (required for OAuth2PasswordRequestForm)
python-multipart==0.0.20

# Database ORM
sqlalchemy>=2.0.44

# Database migrations
alembic==1.12.1

# HTTP requests (you're already using this in api_client.py)
requests==2.32.3

# Environment variable management
python-dotenv==1.0.1

# Pydantic (comes with FastAPI, but pinning version for clarity)
pydantic==2.10.3
pydantic-settings==2.6.1

# Password hashing
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Rate limiting
slowapi==0.1.9

# Testing dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2